+++ qldpc/circuits/__init__.py
@@ -6,7 +6,7 @@ from .common import (
 )
 from .memory import memory_experiment
 from .noise_model import DepolarizingNoiseModel, NoiseModel, NoiseRule, SI1000NoiseModel
-from .sinter_decoders import CompiledSinterDecoder, DetectorErrorModelArrays, SinterDecoder
+from .sinter_decoder import CompiledSinterDecoder, DetectorErrorModelArrays, SinterDecoder
 from .syndrome_measurement import BareColorCircuit, SyndromeMeasurementStrategy
 from .transversal import (
     get_transversal_automorphism_group,
